requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
legacy-cgi==2.6.3
python-dotenv==1.1.0
